+++ src/app/pages/my-requests/my-requests.component.spec.ts
@@ -125,7 +125,6 @@ describe('MyRequestsComponent', () => {
     expect(harness.component.statusFilters).toStrictEqual([]);
     expect(harness.component.request).toBeUndefined();
     expect(harness.component.requestElements).toStrictEqual([]);
-    expect(harness.component.state).toBe('idle');
   });
 
   describe('initialisation', () => {
@@ -137,225 +136,6 @@ describe('MyRequestsComponent', () => {
     it('should do nothing if there is no user', () => {
       harness.component.ngOnInit();
       expect(dataRequestsStub.list).not.toHaveBeenCalled();
-    });
-
-    it('should display all requests available to a user', () => {
-      const requests: DataRequest[] = [
-        {
-          id: '1',
-          label: 'request 1',
-          domainType: CatalogueItemDomainType.DataModel,
-          status: 'unsent',
-        },
-        {
-          id: '2',
-          label: 'request 2',
-          domainType: CatalogueItemDomainType.DataModel,
-          status: 'unsent',
-        },
-      ];
-
-      mockSignedInUser();
-
-      dataRequestsStub.list.mockImplementationOnce(() => {
-        expect(harness.component.state).toBe('loading');
-        return of(requests);
-      });
-
-      harness.component.ngOnInit();
-
-      expect(harness.component.state).toBe('idle');
-      expect(harness.component.allRequests).toStrictEqual(requests);
-      expect(harness.component.filteredRequests).toStrictEqual(requests); // No filters yet
-      expect(harness.component.request).toBe(requests[0]);
-    });
-
-    it('should display an error if failed to get requests', () => {
-      mockSignedInUser();
-
-      dataRequestsStub.list.mockImplementationOnce(() => throwError(() => new Error()));
-
-      harness.component.ngOnInit();
-
-      expect(harness.component.state).toBe('idle');
-      expect(toastrStub.error).toHaveBeenCalled();
-      expect(harness.component.allRequests).toStrictEqual([]);
-      expect(harness.component.filteredRequests).toStrictEqual([]);
-      expect(harness.component.request).toBeUndefined();
-    });
-
-    it('should handle having no requests available', () => {
-      mockSignedInUser();
-
-      dataRequestsStub.list.mockImplementationOnce(() => {
-        expect(harness.component.state).toBe('loading');
-        return of([]);
-      });
-
-      harness.component.ngOnInit();
-
-      expect(harness.component.state).toBe('idle');
-      expect(harness.component.allRequests).toStrictEqual([]);
-      expect(harness.component.filteredRequests).toStrictEqual([]); // No filters yet
-      expect(harness.component.request).toBeUndefined();
-    });
-  });
-
-  describe('has multiple request status', () => {
-    it.each([
-      [false, 0, 0],
-      [false, 1, 0],
-      [false, 0, 1],
-      [true, 2, 2],
-    ])(
-      'should say hasMultipleRequestStatus is %p when unsent count is %p and submitted count is %p',
-      (expected, unsentCount, submittedCount) => {
-        const generateRequests = (count: number, status: DataRequestStatus) => {
-          return [...Array(count).keys()].map(() => {
-            return { status } as DataRequest;
-          });
-        };
-
-        const requests = [
-          ...generateRequests(unsentCount, 'unsent'),
-          ...generateRequests(submittedCount, 'submitted'),
-        ];
-
-        harness.component.allRequests = requests;
-        const actual = harness.component.hasMultipleRequestStatus;
-        expect(actual).toBe(expected);
-      }
-    );
-  });
-
-  describe('select request', () => {
-    beforeEach(() => {
-      toastrStub.error.mockClear();
-    });
-
-    it('should do nothing if no request was selected', () => {
-      const event = {
-        options: [
-          {
-            value: undefined,
-          },
-        ],
-      } as MatSelectionListChange;
-
-      harness.component.selectRequest(event);
-
-      expect(harness.component.state).toBe('idle');
-      expect(harness.component.requestElements).toStrictEqual([]);
-    });
-
-    it('should select the chosen request', () => {
-      const elements: DataElementDto[] = [
-        {
-          id: '1',
-          label: 'element 1',
-          domainType: CatalogueItemDomainType.DataElement,
-        },
-        {
-          id: '2',
-          label: 'element 2',
-          domainType: CatalogueItemDomainType.DataElement,
-        },
-      ];
-
-      const expectedElements: DataElementSearchResult[] = elements.map((e) => {
-        return {
-          ...e,
-          isBookmarked: false,
-          isSelected: false,
-        } as DataElementSearchResult;
-      });
-
-      const request = { id: '1', status: 'unsent' } as DataRequest;
-
-      dataRequestsStub.listDataElements.mockImplementationOnce((req) => {
-        expect(req).toStrictEqual(request);
-        expect(harness.component.state).toBe('loading');
-        return of(elements);
-      });
-
-      let intersections = {} as DataAccessRequestsSourceTargetIntersections;
-
-      dataRequestsStub.getRequestsIntersections.mockImplementationOnce(
-        (model: string) => {
-          intersections = {
-            dataAccessRequests: [
-              {
-                label: 'data model',
-                id: model,
-                domainType: CatalogueItemDomainType.DataModel,
-              },
-            ],
-
-            sourceTargetIntersections: [],
-          };
-          return of(intersections);
-        }
-      );
-
-      explorerStub.getRootDataModel.mockImplementation(() => {
-        return of({
-          label: 'root model',
-          id: 'root model id',
-          domainType: CatalogueItemDomainType.DataModel,
-          finalised: false,
-          availableActions: [],
-        });
-      });
-
-      dataModelsStub.elementsInAnotherModel.mockImplementationOnce(
-        (model, innerElements) => of(innerElements)
-      );
-      dataModelsStub.dataElementToBasic.mockImplementation((element) => {
-        return {
-          id: element.id ?? '',
-          dataClass: element.dataClass ?? '',
-          model: element.model ?? '',
-          label: element.label,
-          isBookmarked: false,
-          breadcrumbs: element.breadcrumbs,
-        };
-      });
-
-      const event = {
-        options: [
-          {
-            value: request,
-          },
-        ],
-      } as MatSelectionListChange;
-
-      harness.component.selectRequest(event);
-
-      expect(harness.component.state).toBe('idle');
-      expect(harness.component.requestElements).toStrictEqual(expectedElements);
-      expect(harness.component.sourceTargetIntersections).toStrictEqual(intersections);
-    });
-
-    it('should display an error if request elements cannot be found', () => {
-      const request = { id: '1', status: 'unsent' } as DataRequest;
-
-      dataRequestsStub.listDataElements.mockImplementationOnce(() =>
-        throwError(() => new Error())
-      );
-
-      const event = {
-        options: [
-          {
-            value: request,
-          },
-        ],
-      } as MatSelectionListChange;
-
-      harness.component.selectRequest(event);
-
-      expect(harness.component.state).toBe('idle');
-      expect(toastrStub.error).toHaveBeenCalled();
-      expect(harness.component.requestElements).toStrictEqual([]);
     });
   });
 
@@ -433,80 +213,6 @@ describe('MyRequestsComponent', () => {
     );
   });
 
-  describe('submit requests', () => {
-    const request: DataRequest = {
-      id: '123',
-      label: 'request',
-      domainType: CatalogueItemDomainType.DataModel,
-      status: 'unsent',
-    };
-
-    beforeEach(() => {
-      researchPluginStub.submitRequest.mockClear();
-      toastrStub.error.mockClear();
-      broadcastStub.dispatch.mockClear();
-      broadcastStub.loading.mockClear();
-    });
-
-    it('should do nothing if there is no request', () => {
-      harness.component.submitRequest();
-
-      expect(researchPluginStub.submitRequest).not.toHaveBeenCalled();
-    });
-
-    it('should do nothing if current request is not in unsent state', () => {
-      harness.component.request = {
-        ...request,
-        status: 'submitted',
-      };
-
-      harness.component.submitRequest();
-
-      expect(researchPluginStub.submitRequest).not.toHaveBeenCalled();
-    });
-
-    it('should raise error if failed to submit', () => {
-      researchPluginStub.submitRequest.mockImplementationOnce((id) => {
-        expect(id).toBe(request.id);
-        return throwError(() => new Error());
-      });
-
-      harness.component.request = request;
-      harness.component.submitRequest();
-
-      expect(researchPluginStub.submitRequest).toHaveBeenCalled();
-      expect(toastrStub.error).toHaveBeenCalled();
-    });
-
-    it('should update status of current request once submitted', () => {
-      const submittedDataModel: DataModelDetail = {
-        id: request.id,
-        label: request.label,
-        domainType: CatalogueItemDomainType.DataModel,
-        availableActions: ['show'],
-        finalised: true,
-        modelVersion: '1.0.0',
-      };
-
-      researchPluginStub.submitRequest.mockImplementationOnce((id) => {
-        expect(id).toBe(request.id);
-        return of(submittedDataModel);
-      });
-
-      harness.component.request = request;
-      harness.component.submitRequest();
-
-      expect(researchPluginStub.submitRequest).toHaveBeenCalled();
-      expect(harness.component.request.status).toBe('submitted');
-      expect(broadcastStub.dispatch).toHaveBeenCalledWith('data-request-submitted');
-      expect(broadcastStub.loading).toHaveBeenCalledWith({
-        isLoading: true,
-        caption: 'Submitting your request...',
-      });
-      expect(broadcastStub.loading).toHaveBeenCalledWith({ isLoading: false });
-    });
-  });
-
   describe('copy request', () => {
     const request: DataRequest = {
       id: '123',
@@ -515,429 +221,5 @@ describe('MyRequestsComponent', () => {
       status: 'submitted',
       modelVersion: '1.0.0',
     };
-
-    beforeEach(() => {
-      dataModelsStub.createFork.mockClear();
-      toastrStub.error.mockClear();
-      broadcastStub.loading.mockClear();
-    });
-
-    it('should do nothing if there is no request', () => {
-      harness.component.copyRequest();
-
-      expect(dataModelsStub.createFork).not.toHaveBeenCalled();
-    });
-
-    it('should do nothing if current request is not in submitted state', () => {
-      harness.component.request = {
-        ...request,
-        status: 'unsent',
-      };
-
-      harness.component.copyRequest();
-
-      expect(dataModelsStub.createFork).not.toHaveBeenCalled();
-    });
-
-    it('should do nothing if current request is not finalised', () => {
-      harness.component.request = {
-        ...request,
-        modelVersion: undefined,
-      };
-
-      harness.component.copyRequest();
-
-      expect(dataModelsStub.createFork).not.toHaveBeenCalled();
-    });
-
-    it('should do nothing if create request dialog was cancelled', () => {
-      dialogsStub.usage.afterClosed.mockImplementationOnce(() => of(undefined));
-
-      harness.component.request = request;
-      harness.component.copyRequest();
-
-      expect(dataModelsStub.createFork).not.toHaveBeenCalled();
-    });
-
-    it('should raise error if failed to create', () => {
-      dialogsStub.usage.afterClosed.mockImplementationOnce(() =>
-        of({ name: 'new request' })
-      );
-
-      dataModelsStub.createFork.mockImplementationOnce((model) => {
-        expect(model.id).toBe(request.id);
-        return throwError(() => new Error());
-      });
-
-      harness.component.request = request;
-      harness.component.copyRequest();
-
-      expect(dataModelsStub.createFork).toHaveBeenCalled();
-      expect(toastrStub.error).toHaveBeenCalled();
-    });
-
-    it('should create new draft copy of request', fakeAsync(() => {
-      const nextDataModel: DataModel = {
-        ...request,
-        modelVersion: undefined,
-      };
-
-      mockSignedInUser();
-
-      dialogsStub.usage.afterClosed.mockImplementationOnce(() =>
-        of({ name: 'new request' })
-      );
-
-      dataModelsStub.createFork.mockImplementationOnce((model) => {
-        expect(model.id).toBe(request.id);
-        return of(nextDataModel);
-      });
-
-      dataRequestsStub.list.mockImplementationOnce(() => of([request, nextDataModel]));
-      dataRequestsStub.listDataElements.mockImplementationOnce(() => of([]));
-
-      harness.component.request = request;
-      harness.component.copyRequest();
-
-      tick();
-
-      expect(dataModelsStub.createFork).toHaveBeenCalled();
-      expect(harness.component.request.status).toBe('unsent');
-      expect(broadcastStub.loading).toHaveBeenCalledWith({
-        isLoading: true,
-        caption: 'Creating new request ...',
-      });
-      expect(broadcastStub.loading).toHaveBeenCalledWith({ isLoading: false });
-    }));
-  });
-
-  describe('remove elements from requests', () => {
-    let elements: DataElementDto[];
-    let selectableElements: () => DataElementSearchResult[];
-    let request: DataRequest;
-    let requestMenuItem = {
-      dataModel: { id: '', label: '', domainType: CatalogueItemDomainType.DataModel },
-      containsElement: false,
-    };
-    let intersections = {} as DataAccessRequestsSourceTargetIntersections;
-    let matSelectionListChangeevent: MatSelectionListChange;
-    let component: MyRequestsComponent;
-    let dom: DebugElement;
-
-    beforeEach(() => {
-      mockSignedInUser();
-      elements = [
-        {
-          id: '1',
-          label: 'element 1',
-          domainType: CatalogueItemDomainType.DataElement,
-        },
-        {
-          id: '2',
-          label: 'element 2',
-          domainType: CatalogueItemDomainType.DataElement,
-        },
-      ];
-
-      selectableElements = () => {
-        return elements.map((e) => {
-          return {
-            ...e,
-            isSelected: false,
-            isBookmarked: false,
-          } as DataElementSearchResult;
-        });
-      };
-
-      request = { id: '1', status: 'unsent', label: 'request 1' } as DataRequest;
-      requestMenuItem = {
-        dataModel: {
-          id: '1',
-          label: 'request 1',
-          domainType: CatalogueItemDomainType.DataModel,
-        },
-        containsElement: true,
-      };
-
-      dataRequestsStub.listDataElements.mockImplementation((req) => {
-        expect(req).toStrictEqual(request);
-        expect(component.state).toBe('loading');
-        return of(elements);
-      });
-
-      dataRequestsStub.list.mockImplementation(() => of([request]));
-
-      dataRequestsStub.getRequestsIntersections.mockImplementation((model: string) => {
-        intersections = {
-          dataAccessRequests: [
-            {
-              label: 'data model',
-              id: model,
-              domainType: CatalogueItemDomainType.DataModel,
-            },
-          ],
-
-          sourceTargetIntersections: [],
-        };
-        return of(intersections);
-      });
-
-      explorerStub.getRootDataModel.mockImplementation(() => {
-        return of({
-          label: 'root model',
-          id: 'root model id',
-          domainType: CatalogueItemDomainType.DataModel,
-          finalised: false,
-          availableActions: [],
-        });
-      });
-
-      dataModelsStub.elementsInAnotherModel.mockImplementation((model, innerElements) =>
-        of(innerElements)
-      );
-      dataModelsStub.dataElementToBasic.mockImplementation((element) => {
-        return {
-          id: element.id ?? '',
-          dataClass: element.dataClass ?? '',
-          model: element.model ?? '',
-          label: element.label,
-          isBookmarked: false,
-          breadcrumbs: element.breadcrumbs,
-        };
-      });
-
-      matSelectionListChangeevent = {
-        options: [
-          {
-            value: request,
-          },
-        ],
-      } as MatSelectionListChange;
-
-      component = harness.component;
-      component.ngOnInit();
-      component.selectRequest(matSelectionListChangeevent);
-      harness.detectChanges();
-      dom = harness.fixture.debugElement;
-    });
-
-    it('Refresh request when data element is deleted via DataElementInRequest component', () => {
-      // pick the first row and trigger the RequestAddDelete event
-      const dataElementRow: DebugElement = dom.query(
-        (de) => de.name === 'mdm-data-element-row'
-      );
-      const requestAddDeleteEvent: RequestElementAddDeleteEvent = {
-        adding: false,
-        dataModel: requestMenuItem.dataModel,
-        dataElement: selectableElements()[0],
-      };
-      // Pretend to delete 1 element
-      dataRequestsStub.listDataElements.mockImplementation((req) => {
-        expect(req).toStrictEqual(request);
-        expect(component.state).toBe('loading');
-        return of(elements.slice(0, 1));
-      });
-      expect(component.requestElements).toStrictEqual(selectableElements());
-      dataElementRow.triggerEventHandler('requestAddDelete', requestAddDeleteEvent);
-      expect(component.requestElements).toStrictEqual(selectableElements().slice(0, 1));
-    });
-
-    it('Don\'t refresh request when data element is deleted from a different request via DataElementInRequest component', () => {
-      // pick the first row and trigger the RequestAddDelete event
-      const dataElementRow: DebugElement = dom.query(
-        (de) => de.name === 'mdm-data-element-row'
-      );
-
-      requestMenuItem.dataModel.id = 'A different id';
-      const requestAddDeleteEvent: RequestElementAddDeleteEvent = {
-        adding: false,
-        dataModel: requestMenuItem.dataModel,
-        dataElement: selectableElements()[0],
-      };
-      // Pretend to delete 1 element
-      dataRequestsStub.listDataElements.mockImplementation(() => {
-        return of(elements.slice(0, 1));
-      });
-      expect(component.requestElements).toStrictEqual(selectableElements());
-      dataElementRow.triggerEventHandler('requestAddDelete', requestAddDeleteEvent);
-      expect(component.requestElements).toStrictEqual(selectableElements());
-    });
-
-    it('Refresh request when data element is added via DataElementInRequest component', () => {
-      // pick the first row and trigger the RequestAddDelete event
-      const dataElementRow: DebugElement = dom.query(
-        (de) => de.name === 'mdm-data-element-row'
-      );
-      requestMenuItem.containsElement = false;
-      const requestAddDeleteEvent: RequestElementAddDeleteEvent = {
-        adding: true,
-        dataModel: requestMenuItem.dataModel,
-        dataElement: selectableElements()[0],
-      };
-      // Pretend to add 1 element
-      elements.push({
-        id: '3',
-        label: 'element 3',
-        domainType: CatalogueItemDomainType.DataElement,
-      });
-      dataRequestsStub.listDataElements.mockImplementation((req) => {
-        expect(req).toStrictEqual(request);
-        expect(component.state).toBe('loading');
-        return of(elements);
-      });
-
-      // expected elements are the original first 2 request elements
-      expect(component.requestElements).toStrictEqual(selectableElements().slice(0, 2));
-      dataElementRow.triggerEventHandler('requestAddDelete', requestAddDeleteEvent);
-      // after refresh, expected elements are all 3 of the request elements
-      expect(component.requestElements).toStrictEqual(selectableElements());
-    });
-
-    it('Should select all data elements when Select All is clicked', () => {
-      // Find the Select All checkbox
-      const matCheckbox: DebugElement = dom.query(
-        (de) => de.name === 'mat-checkbox' && de.nativeElement.innerHTML === 'Select all'
-      );
-      const matCheckboxChange: MatCheckboxChange = {
-        source: {} as MatCheckbox,
-        checked: true,
-      };
-      // Should already be false, but this future-proofs against changes to setup
-      component.requestElements.forEach((element) => (element.isSelected = false));
-      matCheckbox.triggerEventHandler('change', matCheckboxChange);
-      expect(component.requestElements).toHaveLength(2);
-      expect(component.requestElements[0].isSelected).toBe(true);
-      expect(component.requestElements[1].isSelected).toBe(true);
-    });
-
-    it('Should enable RemoveSelected button on when there is a request, there are request elements and at least 1 is selected', () => {
-      const savRequest = component.request;
-      const savRequestElements = component.requestElements;
-
-      component.request = undefined;
-      component.requestElements = [];
-
-      expect(component.removeSelectedButtonDisabled).toBeTruthy();
-
-      component.request = savRequest;
-      expect(component.removeSelectedButtonDisabled).toBeTruthy();
-      component.requestElements = savRequestElements;
-      expect(component.removeSelectedButtonDisabled).toBeTruthy();
-      component.onSelectElement({ item: component.requestElements[0], checked: true });
-      expect(component.removeSelectedButtonDisabled).toBeFalsy();
-      component.onSelectElement({ item: component.requestElements[1], checked: true });
-      expect(component.removeSelectedButtonDisabled).toBeFalsy();
-      component.onSelectElement({ item: component.requestElements[0], checked: false });
-      expect(component.removeSelectedButtonDisabled).toBeFalsy();
-      component.onSelectElement({ item: component.requestElements[1], checked: false });
-      expect(component.removeSelectedButtonDisabled).toBeTruthy();
-    });
-
-    it('Should open OK/Cancel dialogue, show spinner, call into DataRequestsService.deleteDataElementMultiple with a single item and refresh the request list when Remove button is clicked', () => {
-      // Find any data element row from which to trigger an event
-      const dataElementRow: DebugElement = dom.query(
-        (de) => de.name === 'mdm-data-element-row'
-      );
-      const dataElementDeleteEvent: DataElementDeleteEvent = {
-        item: selectableElements()[0],
-      };
-      // spy on OK/Cancel, broadcast service .loading and DataRequestsService.deleteDataElementMultiple
-      dialogsStub.usage.afterClosed.mockReset();
-      const okCancelResponse: OkCancelDialogResponse = {
-        result: true,
-      };
-      dialogsStub.open = jest.fn().mockReturnValue(dialogsStub.usage);
-      dialogsStub.usage.afterClosed.mockReturnValue(of(okCancelResponse));
-      const deleteResult: DataElementMultipleOperationResult = {
-        successes: [
-          {
-            success: true,
-            message: '',
-            item: component.requestElements[0],
-          },
-        ],
-        failures: [],
-      };
-      dataRequestsStub.deleteDataElementMultiple.mockReset();
-      dataRequestsStub.deleteDataElementMultiple.mockReturnValue(of(deleteResult));
-      broadcastStub.loading.mockReset();
-
-      // Pretend to delete first element
-      dataRequestsStub.listDataElements.mockImplementation(() => {
-        return of(elements.slice(1, 2));
-      });
-      // Raise the delete event
-      dataElementRow.triggerEventHandler('delete', dataElementDeleteEvent);
-
-      // check the fallout
-      expect(dialogsStub.open).toHaveBeenCalledTimes(1);
-      expect(broadcastStub.loading).toHaveBeenCalledTimes(2);
-      expect(dataRequestsStub.deleteDataElementMultiple).toHaveBeenCalledTimes(1);
-      expect(dataRequestsStub.deleteDataElementMultiple.mock.calls[0][0]).toStrictEqual([
-        selectableElements()[0],
-      ]);
-      expect(dataRequestsStub.deleteDataElementMultiple.mock.calls[0][1].id).toBe(
-        request.id
-      );
-      expect(dataRequestsStub.deleteDataElementMultiple.mock.calls[0][1].label).toBe(
-        request.label
-      );
-      expect(component.requestElements.length).toBe(1);
-      expect(component.requestElements[0]).toStrictEqual(selectableElements()[1]);
-    });
-
-    it('Should open OK/Cancel dialogue, show spinner, call into DataRequestsService.deleteDataElementMultiple with 2 items and refresh the request list when Remove Selected is clicked', () => {
-      // Find Remove Selected button to click for event trigger
-      const dataElementRow: DebugElement = dom.query(
-        (de) =>
-          de.name === 'button' && de.nativeElement.innerHTML === ' Remove selected ... '
-      );
-      // Select request elements
-      component.requestElements.forEach((el) => (el.isSelected = true));
-      // spy on OK/Cancel, broadcast service .loading and DataRequestsService.deleteDataElementMultiple
-      dialogsStub.usage.afterClosed.mockReset();
-      const okCancelResponse: OkCancelDialogResponse = {
-        result: true,
-      };
-      dialogsStub.open = jest.fn().mockReturnValue(dialogsStub.usage);
-      dialogsStub.usage.afterClosed.mockReturnValue(of(okCancelResponse));
-      const deleteResult: DataElementMultipleOperationResult = {
-        successes: [
-          {
-            success: true,
-            message: '',
-            item: component.requestElements[0],
-          },
-        ],
-        failures: [],
-      };
-      dataRequestsStub.deleteDataElementMultiple.mockReset();
-      dataRequestsStub.deleteDataElementMultiple.mockReturnValue(of(deleteResult));
-      broadcastStub.loading.mockReset();
-
-      // Pretend to delete all elements
-      dataRequestsStub.listDataElements.mockImplementation(() => {
-        return of([]);
-      });
-      // Raise the click event
-      dataElementRow.triggerEventHandler('click', {});
-
-      // check the fallout
-      expect(dialogsStub.open).toHaveBeenCalledTimes(1);
-      expect(broadcastStub.loading).toHaveBeenCalledTimes(2);
-      expect(dataRequestsStub.deleteDataElementMultiple).toHaveBeenCalledTimes(1);
-      const expectedDeleteItems = selectableElements();
-      expectedDeleteItems.forEach((el) => (el.isSelected = true));
-      expect(dataRequestsStub.deleteDataElementMultiple.mock.calls[0][0]).toStrictEqual(
-        expectedDeleteItems
-      );
-      expect(dataRequestsStub.deleteDataElementMultiple.mock.calls[0][1].id).toBe(
-        request.id
-      );
-      expect(dataRequestsStub.deleteDataElementMultiple.mock.calls[0][1].label).toBe(
-        request.label
-      );
-      expect(component.requestElements.length).toBe(0);
-    });
   });
 });
